Switch int to int64 to match msgpack

+++ server/marine.go
@@ -4,8 +4,8 @@ func makeMarine() IUnit {
 	return &unit{
 		id:        newId(),
 		name:      "Marine",
-		health:    70,
-		maxHealth: 70,
+		health:    int64(70),
+		maxHealth: int64(70),
 		tags:      TAG_BIO | TAG_RANGED,
 	}
 }
