+++ server/unit.go
@@ -5,37 +5,37 @@ import (
 )
 
 type unit struct {
-	id        int
+	id        int64
 	name      string
-	health    int
-	maxHealth int
-	tags      int
+	health    int64
+	maxHealth int64
+	tags      int64
 
 	x float64
 	y float64
 }
 
 // getters
-func (u *unit) Id() int        { return u.id }
-func (u *unit) Name() string   { return u.name }
-func (u *unit) Health() int    { return u.health }
-func (u *unit) MaxHealth() int { return u.maxHealth }
-func (u *unit) Tags() int      { return u.tags }
+func (u *unit) Id() int64        { return u.id }
+func (u *unit) Name() string     { return u.name }
+func (u *unit) Health() int64    { return u.health }
+func (u *unit) MaxHealth() int64 { return u.maxHealth }
+func (u *unit) Tags() int64      { return u.tags }
 
 // functionality
-func (u *unit) AddTags(tags ...int) {
+func (u *unit) AddTags(tags ...int64) {
 	for _, tag := range tags {
 		u.tags &= tag
 	}
 }
 
-func (u *unit) RemoveTags(tags ...int) {
+func (u *unit) RemoveTags(tags ...int64) {
 	for _, tag := range tags {
 		u.tags ^= tag
 	}
 }
 
-func (u *unit) HasTags(tags ...int) bool {
+func (u *unit) HasTags(tags ...int64) bool {
 	for _, tag := range tags {
 		if u.tags&tag == 0 {
 			return false
@@ -58,7 +58,7 @@ func (u *unit) Distance(other IUnit) float64 {
 	return math.Hypot(u.x-otherX, u.y-otherY)
 }
 
-func (u *unit) TakeDamage(amount int) (alive bool) {
+func (u *unit) TakeDamage(amount int64) (alive bool) {
 	u.health -= amount
 	alive = u.health > 0
 	return
